[Large Update]
Removed XML Editor.
Removed Idea of JSON Editor.
Changed use of variables on string to use `$` on some files.
Improved File Actions.cs
Created SteamCMD.cs for managing SteamCMD installation at all moments.
Improved Updater.cs in accordance to SteamCMD.cs
Removed SteamLogin class due to security reasons and made steam Username and Password Internally accessible only.
Highly advanced on Workshop Dowloader (NOT FINISHED)

+++ src/File Control/File Actions.cs
@@ -6,8 +6,16 @@ namespace ATORTTeam.UnturnedServerManager.FileControl
 {
     public static class FileActions
     {
+        /// <summary>
+        /// Downloads a file into the destination folder.
+        /// </summary>
+        /// <param name="URL">Uniform Resource Locator for the file.</param>
+        /// <param name="Destination">Full path to destination including desired filename.</param>
+        /// <returns>True if no exceptions occur during download, false otherwise.</returns>
         public static bool Download(string URL, string Destination)
         {
+            VerifyPath(Destination, true);
+
             try
             {
                 using (var client = new WebClient())
@@ -21,11 +29,18 @@ namespace ATORTTeam.UnturnedServerManager.FileControl
                 return false;
             }
         }
+
+        /// <summary>
+        /// Extracts all the contents of a .Zip file to its destination, overwriting the files if necessary.
+        /// </summary>
+        /// <param name="Zip">The full path to the zip file, filename and extension included.</param>
+        /// <param name="Destination">The destination folder for all the contents of the .Zip file to be extracted to.</param>
+        /// <returns>False if the file doesn't exist or if an issue occurs during the entire process of extraction (Ignoring issues when overwriting individual files). Returns true otherwise.</returns>
         public static bool Extract(string Zip, string Destination)
         {
             try
             {
-                VerifyPath(Destination);
+                VerifyPath(Destination, true);
 
                 if (!VerifyFile(Zip))
                     return false;
@@ -36,8 +51,7 @@ namespace ATORTTeam.UnturnedServerManager.FileControl
                     {
                         string dir = Path.GetFullPath(Path.Combine(Destination, Entry.FullName));
 
-                        if (!Directory.Exists(Path.GetDirectoryName(dir)))
-                            Directory.CreateDirectory(dir);
+                        VerifyFilePath(dir, true);
 
                         if (Entry.Name != "")
                             Entry.ExtractToFile(Path.GetFullPath(Path.Combine(Destination, Entry.FullName)), true);
@@ -52,12 +66,14 @@ namespace ATORTTeam.UnturnedServerManager.FileControl
             }
         }
         public static bool VerifyFile(string Path) => File.Exists(Path);
-        public static bool VerifyFilePath(string File) => VerifyPath(Path.GetDirectoryName(File));
-        public static bool VerifyPath(string Folder)
+        public static bool VerifyFilePath(string File, bool Create) => VerifyPath(Path.GetDirectoryName(File), Create);
+        public static bool VerifyPath(string Folder, bool Create)
         {
             if (!Directory.Exists(Folder))
             {
-                Directory.CreateDirectory(Folder);
+                if (Create)
+                    Directory.CreateDirectory(Folder);
+
                 return false;
             }
             else
@@ -65,20 +81,26 @@ namespace ATORTTeam.UnturnedServerManager.FileControl
         }
         public static void DeleteDirectory(string Folder)
         {
-            VerifyPath(Folder);
-            try
+            if (VerifyPath(Folder, false))
             {
-                Directory.Delete(Folder, true);
+                try
+                {
+                    Directory.Delete(Folder, true);
+                }
+                catch { }
             }
-            catch { }
         }
         public static void CopyDirectory(string Folder, string Destination)
         {
-            foreach (string dirPath in Directory.GetDirectories(Folder, "*", SearchOption.AllDirectories))
-                Directory.CreateDirectory(dirPath.Replace(Folder, Destination));
-            
-            foreach (string newPath in Directory.GetFiles(Folder, "*.*", SearchOption.AllDirectories))
-                File.Copy(newPath, newPath.Replace(Folder, Destination), true);
+            if (VerifyPath(Folder, false))
+            {
+                VerifyPath(Destination, true);
+                foreach (string dirPath in Directory.GetDirectories(Folder, "*", SearchOption.AllDirectories))
+                    Directory.CreateDirectory(dirPath.Replace(Folder, Destination));
+
+                foreach (string newPath in Directory.GetFiles(Folder, "*.*", SearchOption.AllDirectories))
+                    File.Copy(newPath, newPath.Replace(Folder, Destination), true);
+            }
         }
     }
 }
